+++ spec/features/staff/users_can_manage_activity_geography_spec.rb
@@ -2,7 +2,7 @@ RSpec.feature "Users can provide the geography for an activity" do
   context "when the user belongs to BEIS" do
     let(:user) { create(:beis_user) }
     before { authenticate!(user: user) }
-    let(:activity) { create(:activity, :at_geography_step) }
+    let(:activity) { create(:activity, :at_geography_step, organisation: user.organisation) }
 
     scenario "they are asked to choose the geography" do
       visit activity_step_path(activity, :geography)
@@ -59,7 +59,10 @@ RSpec.feature "Users can provide the geography for an activity" do
 
     context "with a completed activity" do
       scenario "they can change the geography from region to country" do
-        activity = create(:activity, geography: :recipient_region, recipient_country: nil)
+        activity = create(:activity,
+          geography: :recipient_region,
+          recipient_country: nil,
+          organisation: user.organisation)
         activity_path = organisation_activity_details_path(activity.organisation, activity)
 
         visit activity_path
@@ -83,7 +86,11 @@ RSpec.feature "Users can provide the geography for an activity" do
       end
 
       scenario "they can change the geography from country to region" do
-        activity = create(:activity, geography: :recipient_country, recipient_region: nil, recipient_country: "AG")
+        activity = create(:activity,
+          geography: :recipient_country,
+          recipient_region: nil,
+          recipient_country: "AG",
+          organisation: user.organisation)
         activity_path = organisation_activity_details_path(activity.organisation, activity)
 
         visit activity_path