+++ CHANGELOG.md
@@ -202,6 +202,7 @@
 - add accessibility statement
 - Move budgets to the top of the programme activity financials tab
 - Activity XML includes only recipient-country OR recipient-region, not both
+- Activity creation journey changed to ask for a level and parent
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-11...HEAD
 [release-11]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-10...release-11